feat: 기본 UI 구조 구현
